Rename relational input example algorithm.
